64-bit port
Build and run the engine as a 64-bit (LP64) binary; the build system
previously hard-errored on 64-bit targets. Two structural hazards on LP64 are
addressed: pointers no longer fit in the 32-bit fields that UnrealScript
mirrors as ints, and native C++ class layout otherwise diverges from the
script-compiled layout.

Pointer-sized integer types
- Add UPTRINT/PTRINT (uintptr_t/intptr_t) in UnGnuG.h and UnVcWin32.h and use
  them for every pointer<->integer round-trip: FMemStack and Align() alignment
  math, STRUCT_OFFSET, CPP_PROPERTY, UStrProperty binding offsets,
  RandomBase(Owner), and WAV chunk-bounds checks.
- operator new/new[] take size_t; pointer logging uses %p instead of
  %08X/(INT)Ptr.

Script bytecode object references kept at 4 bytes
- Object references embedded in compiled bytecode were raw pointers; at 8 bytes
  they would corrupt every baked jump/skip offset in packages compiled by the
  32-bit UCC. Store them as 4-byte GObjObjects indices instead (XFER_OBJ in
  UStruct::SerializeExpr) and resolve back to pointers at run time via
  UObject::GetIndexedObject (FFrame::ReadObject, execBoolVariable). The on-disk
  package format is unchanged, so .u files stay compatible both ways.

Native class layout matched to the script mirror (which assumes pack(4))
- Apply #pragma pack(4) to the hand-written native class headers, and on GCC
  (not only MSVC) to the generated *Classes.h.
- Widen the script-side placeholders for hidden C++ pointer members
  (GetPlaceholderSize): ObjectInternal, vtable stubs, Canvas FramePtr/RenderPtr,
  ScriptedTexture Junk*, etc. occupy pointer-sized storage; the mirrored int
  fields Context, LogAr and PrivateResolveInfo widen to PTRINT.
- Record each native class's real sizeof as UStruct::IntrinsicSize and warn at
  link time on any script-vs-C++ size mismatch (a correct build logs none);
  IntrinsicSize is preserved across the in-place class reconstruction in
  UObject::StaticAllocateObject.
- GCC_ALIGN(4) the leading BYTE member of AMover/APlayerStart/ATcpLink so it is
  pushed out of the base class tail padding and derived offsets match script.
- Stash native registration info in pointer-sized fields (Outer/StateFrame/
  _Linker) rather than the 4-byte Name/_LinkerIndex, and stash the class config
  name via NativeConfigNameStash.

Desktop (SDL/GL) build support that accompanies the 64-bit desktop target
- CMakeLists: stop forcing 32-bit / erroring on 64-bit builds.
- NSDLViewport: populate the key map in the constructor, enable vsync (an
  uncapped frame rate made game time run several times too fast), and capture
  the mouse on click for windowed mouselook.
- NOpenGLESDrv: detect a desktop GL context at run time and compile the GLES2
  shaders as GLSL 120 (stubbing the ES precision qualifiers).
- Add UNSDLClient / UNOpenGLESRenderDevice StaticConstructor hooks.
- Include SDL before the engine headers so UnFile.h's clock(Timer) macro does
  not mangle time.h's clock(); raise MAX_NODES/MAX_POINTS on non-Dreamcast
  targets.

+++ Source/Core/Inc/UnFile.h
@@ -450,13 +450,13 @@ inline void* appRealloc( void* Original, DWORD Count, const TCHAR* Tag )
 //
 // C++ style memory allocation.
 //
-inline void* operator new( unsigned int Size, const TCHAR* Tag )
+inline void* operator new( size_t Size, const TCHAR* Tag )
 {
 	guardSlow(new);
 	return appMalloc( Size, Tag );
 	unguardSlow;
 }
-inline void* operator new( unsigned int Size )
+inline void* operator new( size_t Size )
 {
 	guardSlow(new);
 	return appMalloc( Size, TEXT("new") );
@@ -470,13 +470,13 @@ inline void operator delete( void* Ptr )
 }
 
 #if PLATFORM_NEEDS_ARRAY_NEW
-inline void* operator new[]( unsigned int Size, const TCHAR* Tag )
+inline void* operator new[]( size_t Size, const TCHAR* Tag )
 {
 	guardSlow(new);
 	return appMalloc( Size, Tag );
 	unguardSlow;
 }
-inline void* operator new[]( unsigned int Size )
+inline void* operator new[]( size_t Size )
 {
 	guardSlow(new);
 	return appMalloc( Size, TEXT("new") );
